Split `shard_holder.rs` into submodules (#4561)
* Move `collection/src/shards/shard_holder.rs` to `collection/src/shards/shard_holder/mod.rs`

* Move `ShardHolder`'s resharding-related methods into `resharding` submodule
